# swagger/swagger.json
{"swagger":"2.0","info":{"title":"The Hello App","description":"Say Hello","version":""},"host":"localhost:8080","schemes":["http"],"consumes":["application/json","application/xml","application/gob","application/x-gob"],"produces":["application/json","application/xml","application/gob","application/x-gob"],"paths":{"/hello/{name}":{"get":{"tags":["Hello"],"summary":"hello Hello","description":"API calls your name","operationId":"Hello#hello","parameters":[{"name":"name","in":"path","description":"Name","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/GoaExampleHello"}}},"schemes":["http"]}},"/swagger.json":{"get":{"summary":"Download swagger/swagger.json","operationId":"swagger#/swagger.json","responses":{"200":{"description":"File downloaded","schema":{"type":"file"}}},"schemes":["http"]}},"/swaggerui/{filepath}":{"get":{"summary":"Download swaggerui/dist","operationId":"swagger#/swaggerui/*filepath","parameters":[{"name":"filepath","in":"path","description":"Relative file path","required":true,"type":"string"}],"responses":{"200":{"description":"File downloaded","schema":{"type":"file"}},"404":{"description":"File not found","schema":{"$ref":"#/definitions/error"}}},"schemes":["http"]}}},"definitions":{"GoaExampleHello":{"title":"Mediatype identifier: application/vnd.goa.example.hello+json; view=default","type":"object","properties":{"msg":{"type":"string","example":"Quia quia labore magni numquam aut."}},"description":"Message from Demo App API (default view)","example":{"msg":"Quia quia labore magni numquam aut."},"required":["msg"]},"error":{"title":"Mediatype identifier: application/vnd.goa.error; view=default","type":"object","properties":{"code":{"type":"string","description":"an application-specific error code, expressed as a string value.","example":"invalid_value"},"detail":{"type":"string","description":"a human-readable explanation specific to this occurrence of the problem.","example":"Value of ID must be an integer"},"id":{"type":"string","description":"a unique identifier for this particular occurrence of the problem.","example":"3F1FKVRR"},"meta":{"type":"object","description":"a meta object containing non-standard meta-information about the error.","example":{"timestamp":1458609066},"additionalProperties":true},"status":{"type":"string","description":"the HTTP status code applicable to this problem, expressed as a string value.","example":"400"}},"description":"Error response media type (default view)","example":{"code":"invalid_value","detail":"Value of ID must be an integer","id":"3F1FKVRR","meta":{"timestamp":1458609066},"status":"400"}}},"responses":{"OK":{"description":"OK","schema":{"$ref":"#/definitions/GoaExampleHello"}}}}